更新项目

+++ FourOperations/FourOperations/BaseData/BasicData.cs
@@ -4,13 +4,13 @@ namespace ConsoleApp2
 {
     public abstract class BasicData : IBasicData
     {
-        public int BasicDataA { get; set; }
-        public int BasicDataB { get; set; }
-        public int BasicDataC { get; set; }
+        public double BasicDataA { get; set; }
+        public double BasicDataB { get; set; }
+        public double BasicDataC { get; set; }
         public int LsitNumA { get; set; }
         public int LsitNumB { get; set; }
         public string Theme { get; set; }//题目
-        public decimal Answer { get; set; }//答案
+        public double Answer { get; set; }//答案
         protected string FFF { get; set; } = "=";
 
         public List<string> CCC = new List<string> { "+", "-", "*", "/" };
